# apps/backend/src/shared/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mandarin Learning Platform API
  version: 1.0.0
  description: |
    Production backend API for Mandarin vocabulary learning and AI-powered
    conversation generation. All endpoints are versioned under /api/v1.

    This reference documents the full NestJS 11 route surface: 54 route paths
    across the 15 backend modules (audio, auth, characters, chengyu,
    foundations, grammar, health, mnemonics, phonetic-clusters, progression,
    quiz, radicals, readers, review, words — pinyin lives under foundations).

    Authentication model (per route):
      - Public — no auth required (static reference data, health, auth
        register/login/refresh/logout).
      - Required auth — `Authorization: Bearer <access_token>`; user-scoped
        state or AI/vendor-cost generation (`RequireAuthGuard`).
      - Optional auth — bearer token is best-effort (`OptionalAuthGuard`): an
        absent or invalid token proceeds as a guest (`req.userId` undefined),
        never 401. Guest responses are session-local/empty shapes.

    All 4xx/5xx errors use the `{ code, message, requestId }` envelope.
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://mandarin-vite-react-ts-production.up.railway.app/api
    description: Production server (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HTTP-only refresh token cookie
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Standard error envelope — the global 24-3 `AppExceptionFilter`
        serializes every thrown `HttpException` into this shape.
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request ID (echoes the `X-Request-Id` response header)
      required:
        - code
        - message
        - requestId
...
paths:
...
  /v1/progression/foundation-progress:
    get:
      summary: Get foundation section progress
      tags:
        - Progression
      description: |
        Fetches the user's foundation section progress (auto-initializes four
        records when none exist). Optional auth — a guest receives `[]`.
      responses:
        "200":
          description: Foundation progress records (empty array for guests)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "500":
          description: Failed to fetch foundation progress (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/progression/foundation-progress/{sectionId}:
    put:
      summary: Mark a foundation section completed
      tags:
        - Progression
      security:
        - bearerAuth: []
      description: Marks a foundation section as completed (required auth).
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
          description: Foundation section identifier
      responses:
        "200":
          description: Updated progress record
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid sectionId (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to mark section completed (UPDATE_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/progression/phase-gate:
    get:
      summary: Get phase gate status
      tags:
        - Progression
      description: |
        Fetches the user's phase gate status (auto-creates with defaults when
        none exists). Optional auth — a guest receives the calibrated guest
        gate (`currentPhase: 1, isGuest: true`).
      responses:
        "200":
          description: Phase gate object
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Failed to fetch phase gate (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      summary: Update phase gate progression
      tags:
        - Progression
      security:
        - bearerAuth: []
      description: Updates phase gate progression after a quiz attempt (required auth).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                phase:
                  type: number
                passed:
                  type: boolean
                gateCriteria:
                  type: string
      responses:
        "200":
          description: Updated phase gate object
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to update phase gate (UPDATE_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/progression/gates:
    get:
      summary: Get computed gate statuses
      tags:
        - Progression
      description: |
        Fetches COMPUTED gate statuses (Phase 2 IME, character count >= 500,
        Phase 3→4 comprehension). Optional auth — a CALIBRATED guest is
        Phase-1-only, so the Phase 2+ gates are not passed.
      responses:
        "200":
          description: Computed gate statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  phase2Gate:
                    type: object
                  characterCountGate:
                    type: object
                  phase3To4Gate:
                    type: object
        "500":
          description: Failed to load gate status (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/progression/radical-progress:
    get:
      summary: Get radical progress
      tags:
        - Progression
      description: |
        Fetches the user's radical progress records. Optional auth — a guest
        receives `[]`.
      responses:
        "200":
          description: Radical progress records (empty array for guests)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "500":
          description: Failed to load radical progress (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/progression/radical-progress/{radicalId}:
    put:
      summary: Create or update radical progress
      tags:
        - Progression
      security:
        - bearerAuth: []
      description: |
        Creates or updates radical progress. When `memorized: true`, a
        ReviewItem side-effect is fired (required auth).
      parameters:
        - name: radicalId
          in: path
          required: true
          schema:
            type: string
          description: Radical business key (e.g. rad_XXXX)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                memorized:
                  type: boolean
                recognitionLevel:
                  type: number
      responses:
        "200":
          description: Updated progress record
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid radicalId (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to update radical progress (UPDATE_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ══════════════════════════════════════════════════════════════════
  # Quiz
  # ══════════════════════════════════════════════════════════════════
...